modularise modal styles
