+++ packages/core-ui/lib/modal/ModalBackground/index.scss
@@ -1,0 +1,9 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-background {
+  @include modal-position;
+
+  background: var(--modal-background-color);
+  z-index: 9;
+}
