+++ packages/core-ui/lib/modal/ModalBackground/index.tsx
@@ -1,0 +1,16 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ModalAnimationProps } from "../types";
+import { motion } from "framer-motion";
+import "./index.scss";
+
+/**
+ * @name ModalBackground
+ * @summary Modal background wrapper.
+ */
+export const ModalBackground = ({ children, ...rest }: ModalAnimationProps) => (
+  <motion.div className="modal-background" {...rest}>
+    {children}
+  </motion.div>
+);
