+++ packages/core-ui/lib/modal/ModalConnectItem/index.scss
@@ -1,0 +1,42 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles/variables" as v;
+
+.modal-connect-item {
+  padding: 0.5rem;
+  flex-grow: 0;
+  width: 50%;
+
+  @media (max-width: v.$modal-two-item-threshold) {
+    width: 100%;
+  }
+
+  &.can-connect {
+    > .inner {
+      transition: transform var(--transition-duration);
+
+      &:hover {
+        transform: scale(1.015);
+      }
+    }
+  }
+
+  .foot {
+    padding: 0.85rem;
+
+    .link {
+      color: var(--text-color-secondary);
+      transition: color var(--transition-duration);
+
+      &:hover {
+        color: var(--network-color-primary);
+      }
+
+      > svg {
+        margin-left: 0.35rem;
+        margin-top: 0.3rem;
+      }
+    }
+  }
+}
