+++ packages/core-ui/lib/modal/ModalContainer/index.scss
@@ -1,0 +1,25 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-container {
+  @include modal-position;
+
+  z-index: 9;
+
+  > div {
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 1rem 2rem;
+
+    /* click anywhere behind modal content to close */
+    .close {
+      position: fixed;
+      width: 100%;
+      height: 100%;
+      z-index: 8;
+      cursor: default;
+    }
+  }
+}
