+++ packages/core-ui/lib/modal/ModalCustomHeader/index.scss
@@ -1,0 +1,43 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-custom-header {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  margin-top: 0.5rem;
+  position: relative;
+
+  > div {
+    &.first {
+      display: flex;
+      flex-grow: 1;
+
+      h1 {
+        color: var(--text-color-primary);
+        font-size: 1.4rem;
+        font-family: Unbounded, sans-serif;
+        font-weight: 700;
+        padding: 0 0.5rem;
+        display: flex;
+        align-items: center;
+
+        > button {
+          margin-left: 1.25rem;
+        }
+      }
+    }
+  }
+
+  .close-modal {
+    position: absolute;
+    top: 0;
+    right: 1.5rem;
+    opacity: 0.25;
+
+    &:hover {
+      opacity: 1;
+    }
+  }
+}
