+++ packages/core-ui/lib/modal/ModalFixedTitle/index.scss
@@ -1,0 +1,13 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-fixed-title {
+  width: 100%;
+  background: transparent;
+  padding: 1rem 0 0;
+
+  &.with-style {
+    background: var(--background-default);
+    padding: 1rem 1rem 0;
+  }
+}
