+++ packages/core-ui/lib/modal/ModalFooter/index.tsx
@@ -1,0 +1,15 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ComponentBase } from "lib/types";
+import "./index.scss";
+
+/**
+ * @name ModalFooter
+ * @summary Used for extrinsics forms.
+ */
+export const ModalFooter = ({ children, style }: ComponentBase) => (
+  <div className="modal-footer" style={style}>
+    {children}
+  </div>
+);
