+++ packages/core-ui/lib/modal/ModalHardwareItem/index.tsx
@@ -1,0 +1,15 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ComponentBase } from "lib/types";
+import "./index.scss";
+
+/**
+ * @name  ModalHardwareItem
+ * @summary Inner wrapper for a hardware connect item.
+ */
+export const ModalHardwareItem = ({ children, style }: ComponentBase) => (
+  <div className={"modal-hardware-item"} style={style}>
+    {children}
+  </div>
+);
