+++ packages/core-ui/lib/modal/ModalHeight/index.scss
@@ -1,0 +1,20 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-height {
+  transition: height 0.5s cubic-bezier(0.1, 1, 0.2, 1);
+  width: 100%;
+  max-width: 600px;
+  max-height: 100%;
+  border-radius: 1.5rem;
+  position: relative;
+  z-index: 9;
+
+  &.large {
+    max-width: 800px;
+  }
+
+  &.xl {
+    max-width: 1250px;
+  }
+}
