+++ packages/core-ui/lib/modal/ModalMotionTwoSection/index.scss
@@ -1,0 +1,8 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-motion-two-sections {
+  @include modal-motion-sections;
+
+  width: 200%;
+}
