+++ packages/core-ui/lib/modal/ModalMotionTwoSection/index.tsx
@@ -1,0 +1,19 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import { ModalAnimationProps } from "../types";
+import { motion } from "framer-motion";
+import "./index.scss";
+
+/**
+ * @name ModalMotionTwoSection
+ * @summary Two section wrapper with motion animation.
+ */
+export const ModalMotionTwoSection = ({
+  children,
+  ...rest
+}: ModalAnimationProps) => (
+  <motion.div className="modal-motion-two-sections" {...rest}>
+    {children}
+  </motion.div>
+);
