+++ packages/core-ui/lib/modal/ModalNotes/index.scss
@@ -1,0 +1,15 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-notes {
+  width: 100%;
+  padding: 0;
+
+  &.with-padding {
+    padding: 0.75rem 0;
+  }
+
+  > p {
+    color: var(--text-color-secondary);
+  }
+}
