+++ packages/core-ui/lib/modal/ModalPadding/index.scss
@@ -1,0 +1,16 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-padding {
+  display: flex;
+  flex-flow: column wrap;
+  padding: 1rem;
+
+  &.vertical-only {
+    padding: 1rem 0 0.25rem;
+  }
+
+  &.horizontal-only {
+    padding: 0 1rem;
+  }
+}
