+++ packages/core-ui/lib/modal/ModalSection/index.scss
@@ -1,0 +1,17 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+
+.modal-carousel {
+  width: 100%;
+  display: flex;
+  flex-flow: column wrap;
+  padding: 0;
+  position: relative;
+}
+
+.modal-tabs {
+  width: 100%;
+  display: flex;
+  margin: 1.25rem 0 0.25rem;
+}
