+++ packages/core-ui/lib/modal/ModalSeparator/index.scss
@@ -1,0 +1,9 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+.modal-separator {
+  border-top: 1px solid var(--text-color-secondary);
+  width: 100%;
+  opacity: 0.1;
+  margin: 0.8rem 0;
+}
