+++ packages/core-ui/lib/modal/ModalSeparator/index.tsx
@@ -1,0 +1,10 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+import "./index.scss";
+
+/**
+ * @name ModalSeparator
+ * @summary A line to separate the content.
+ */
+export const ModalSeparator = () => <div className="modal-separator" />;
