+++ packages/core-ui/lib/modal/styles.scss
@@ -1,8 +1,6 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-@use "../styles/variables" as v;
-
 @mixin modal-position {
   position: fixed;
   width: 100%;
@@ -15,351 +13,4 @@ SPDX-License-Identifier: Apache-2.0 */
   display: flex;
   overflow: auto;
   position: relative;
-}
-
-.modal-footer {
-  display: flex;
-  align-items: center;
-  width: 100%;
-
-  h3 {
-    color: var(--text-color-secondary);
-    opacity: 0.5;
-    position: relative;
-    top: 1.25rem;
-
-    &.active {
-      color: var(--network-color-primary);
-      opacity: 1;
-    }
-  }
-}
-
-.modal-background {
-  @include modal-position;
-
-  background: var(--modal-background-color);
-  z-index: 9;
-}
-
-.modal-container {
-  @include modal-position;
-
-  z-index: 9;
-
-  > div {
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 1rem 2rem;
-
-    /* click anywhere behind modal content to close */
-    .close {
-      position: fixed;
-      width: 100%;
-      height: 100%;
-      z-index: 8;
-      cursor: default;
-    }
-  }
-}
-
-.modal-height {
-  transition: height 0.5s cubic-bezier(0.1, 1, 0.2, 1);
-  width: 100%;
-  max-width: 600px;
-  max-height: 100%;
-  border-radius: 1.5rem;
-  position: relative;
-  z-index: 9;
-
-  &.large {
-    max-width: 800px;
-  }
-
-  &.xl {
-    max-width: 1250px;
-  }
-}
-
-.modal-action-item {
-  border-bottom: 1px solid var(--border-primary-color);
-  color: var(--text-color-primary);
-  font-family: InterSemiBold, sans-serif;
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  margin: 1.25rem 0 0;
-  padding-bottom: 0.75rem;
-  width: 100%;
-
-  > svg {
-    margin-right: 0.65rem;
-  }
-
-  .toggle {
-    background: var(--background-input);
-    border: 1px solid var(--border-primary-color);
-    border-radius: 0.4rem;
-    margin-right: 0.65rem;
-    width: 1.75rem;
-    height: 1.75rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  span {
-    margin-left: 0.5rem;
-  }
-}
-
-.modal-padding {
-  display: flex;
-  flex-flow: column wrap;
-  padding: 1rem;
-
-  &.vertical-only {
-    padding: 1rem 0 0.25rem;
-  }
-
-  &.horizontal-only {
-    padding: 0 1rem;
-  }
-}
-
-.modal-separator {
-  border-top: 1px solid var(--text-color-secondary);
-  width: 100%;
-  opacity: 0.1;
-  margin: 0.8rem 0;
-}
-
-.modal-notes {
-  width: 100%;
-  padding: 0;
-
-  &.with-padding {
-    padding: 0.75rem 0;
-  }
-
-  > p {
-    color: var(--text-color-secondary);
-  }
-}
-
-.modal-warnings {
-  margin-top: 0;
-  width: 100%;
-
-  &.with-margin {
-    margin-top: 0.75rem;
-  }
-}
-
-.modal-custom-header {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  margin-top: 0.5rem;
-  position: relative;
-
-  > div {
-    &.first {
-      display: flex;
-      flex-grow: 1;
-
-      h1 {
-        color: var(--text-color-primary);
-        font-size: 1.4rem;
-        font-family: Unbounded, sans-serif;
-        font-weight: 700;
-        padding: 0 0.5rem;
-        display: flex;
-        align-items: center;
-
-        > button {
-          margin-left: 1.25rem;
-        }
-      }
-    }
-  }
-
-  .close-modal {
-    position: absolute;
-    top: 0;
-    right: 1.5rem;
-    opacity: 0.25;
-
-    &:hover {
-      opacity: 1;
-    }
-  }
-}
-
-.modal-fixed-title {
-  width: 100%;
-  background: transparent;
-  padding: 1rem 0 0;
-
-  &.with-style {
-    background: var(--background-default);
-    padding: 1rem 1rem 0;
-  }
-}
-
-.modal-carousel {
-  width: 100%;
-  display: flex;
-  flex-flow: column wrap;
-  padding: 0;
-  position: relative;
-}
-
-.modal-tabs {
-  width: 100%;
-  display: flex;
-  margin: 1.25rem 0 0.25rem;
-}
-
-.modal-motion-two-sections {
-  @include modal-motion-sections;
-
-  width: 200%;
-}
-
-.modal-motion-three-sections {
-  @include modal-motion-sections;
-
-  width: 300%;
-
-  > .section {
-    border-radius: 1rem;
-    display: flex;
-    flex-direction: column;
-    flex-basis: 33.33%;
-    min-width: 33.33%;
-    height: auto;
-    overflow: auto;
-    flex-grow: 1;
-  }
-}
-
-.modal-content {
-  background: var(--background-modal);
-
-  h2 {
-    margin: 1rem 0;
-
-    &.unbounded {
-      font-family: Unbounded, sans-serif;
-      font-weight: 700;
-    }
-
-    &.title {
-      font-size: 1.35rem;
-      margin: 1.25rem 0 0;
-    }
-  }
-
-  a {
-    color: var(--network-color-primary);
-  }
-
-  .header {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    padding: 1rem 1rem 0;
-  }
-
-  .body {
-    padding: 1rem;
-  }
-}
-
-.modal-connect-item {
-  padding: 0.5rem;
-  flex-grow: 0;
-  width: 50%;
-
-  @media (max-width: v.$modal-two-item-threshold) {
-    width: 100%;
-  }
-
-  &.can-connect {
-    > .inner {
-      transition: transform var(--transition-duration);
-
-      &:hover {
-        transform: scale(1.015);
-      }
-    }
-  }
-
-  .foot {
-    padding: 0.85rem;
-
-    .link {
-      color: var(--text-color-secondary);
-      transition: color var(--transition-duration);
-
-      &:hover {
-        color: var(--network-color-primary);
-      }
-
-      > svg {
-        margin-left: 0.35rem;
-        margin-top: 0.3rem;
-      }
-    }
-  }
-}
-
-.modal-hardware-item {
-  background: var(--button-primary-background);
-  border-radius: 1rem;
-  position: relative;
-  overflow: hidden;
-  width: 100%;
-
-  .body {
-    padding: 1.35rem 0.85rem 0.75rem;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-
-    .row {
-      display: flex;
-      justify-content: center;
-
-      p {
-        margin-bottom: 0;
-      }
-
-      &.margin {
-        margin-top: 0.75rem;
-      }
-
-      .logo {
-        height: 3rem;
-        margin: 0.75rem 0.5rem 0.5rem;
-
-        &.mono {
-          path {
-            fill: var(--text-color-secondary);
-          }
-        }
-      }
-    }
-
-    .status {
-      position: absolute;
-      top: 0.9rem;
-      right: 0.9rem;
-    }
-  }
 }
